multi ner framework & some bug fixed

+++ src/ner/extractor.h
@@ -13,7 +13,7 @@ namespace ner {
 class Extractor {
 public:
   static Extractor& extractor();
-  static int num_templates();
+  static size_t num_templates();
   static int extract1o(const Instance& inst, int idx,
       std::vector<utility::StringVec> & cache);
